Cleanup.

+++ src/c4/test/tstInvert_Comm_Map.cc
@@ -37,21 +37,21 @@ void test2(rtt_c4::ParallelUnitTest &ut) {
   invert_comm_map(to_nodes, from_nodes);
 
   if (node == 0) {
-      if (static_cast<int>(from_nodes.size()) != 0)
-          FAILMSG("Incorrect map size on node 0.");
+    if (from_nodes.size() != 0u)
+      FAILMSG("Incorrect map size on node 0.");
   }
 
   if (node == 1) {
-      if (static_cast<int>(from_nodes.size()) != 1)
-          FAILMSG("Incorrect size of map on node 1.");
-      if (from_nodes[0] != 0)
-          FAILMSG("Incorrect map contents on node 1.");
+    if (from_nodes.size() != 1u)
+      FAILMSG("Incorrect size of map on node 1.");
+    if (from_nodes[0] != 0)
+      FAILMSG("Incorrect map contents on node 1.");
   }
 
   if (ut.numFails == 0)
-      ut.passes("test2 passes");
+    ut.passes("test2 passes");
   else
-      ut.failure("test2 failed");
+    ut.failure("test2 failed");
 
   return;
 }
@@ -62,18 +62,18 @@ void test4(rtt_c4::ParallelUnitTest &ut) {
   std::vector<int> to_nodes;
 
   if (node == 0) {
-      to_nodes.push_back(1);
-      to_nodes.push_back(2);
-      to_nodes.push_back(3);
+    to_nodes.push_back(1);
+    to_nodes.push_back(2);
+    to_nodes.push_back(3);
   }
   if (node == 1) {
-      to_nodes.push_back(0);
+    to_nodes.push_back(0);
   }
   if (node == 2) {
-      to_nodes.push_back(0);
+    to_nodes.push_back(0);
   }
   if (node == 3) {
-      to_nodes.push_back(0);
+    to_nodes.push_back(0);
   }
 
   std::vector<int> from_nodes(0);
@@ -81,23 +81,23 @@ void test4(rtt_c4::ParallelUnitTest &ut) {
   invert_comm_map(to_nodes, from_nodes);
 
   if (node == 0) {
-      if (static_cast<int>(from_nodes.size()) != 3)
-          FAILMSG("Incorrect map size on node 0");
-      for (int i = 0; i < 3; ++i) {
-          if (from_nodes[i] != i+1)
-              FAILMSG("Incorrent map contents on node 0");
+    if (from_nodes.size() != 3u)
+      FAILMSG("Incorrect map size on node 0");
+    for (int i = 0; i < 3; ++i) {
+      if (from_nodes[i] != i + 1)
+        FAILMSG("Incorrent map contents on node 0");
     }
   } else {
-      if (static_cast<int>(from_nodes.size()) != 1)
-          FAILMSG("Incorrect map size.");
-      if (from_nodes[0] != 0)
-          FAILMSG("Incorrect map contents.");
+    if (from_nodes.size() != 1u)
+      FAILMSG("Incorrect map size.");
+    if (from_nodes[0] != 0)
+      FAILMSG("Incorrect map contents.");
   }
 
   if (ut.numFails == 0)
-      ut.passes("test4 passes");
+    ut.passes("test4 passes");
   else
-      ut.failure("test4 failed");
+    ut.failure("test4 failed");
 
   return;
 }
@@ -115,7 +115,7 @@ void test_n_to_n(rtt_c4::ParallelUnitTest &ut) {
   invert_comm_map(to_nodes, from_nodes);
 
   if (static_cast<int>(from_nodes.size()) != nodes)
-      FAILMSG("Incorrect from_nodes size.");
+    FAILMSG("Incorrect from_nodes size.");
 
   for (int i = 0; i < nodes; ++i) {
     if (to_nodes[0] != from_nodes[0])
@@ -123,9 +123,9 @@ void test_n_to_n(rtt_c4::ParallelUnitTest &ut) {
   }
 
   if (ut.numFails == 0)
-      ut.passes("test_n_to_n passes");
+    ut.passes("test_n_to_n passes");
   else
-      ut.failure("test_n_to_n failed");
+    ut.failure("test_n_to_n failed");
 
   return;
 }
@@ -142,15 +142,15 @@ void test_cyclic(rtt_c4::ParallelUnitTest &ut) {
   std::vector<int> from_nodes;
   invert_comm_map(to_nodes, from_nodes);
 
-  if (static_cast<int>(from_nodes.size()) != 1)
+  if (from_nodes.size() != 1u)
     FAILMSG("Incorrect map size.");
   if (from_nodes[0] != (node + nodes - 1) % nodes)
     FAILMSG("Incorrect map contents in cyclc test.");
 
   if (ut.numFails == 0)
-      ut.passes("test_cyclic passes");
+    ut.passes("test_cyclic passes");
   else
-      ut.failure("test_cyclic failed");
+    ut.failure("test_cyclic failed");
 
   return;
 }
@@ -163,13 +163,13 @@ void test_empty(rtt_c4::ParallelUnitTest &ut) {
 
   invert_comm_map(to_nodes, from_nodes);
 
-  if (static_cast<int>(from_nodes.size()) != 0)
+  if (from_nodes.size() != 0u)
     FAILMSG("Incorrect map size in empty test.");
 
   if (ut.numFails == 0)
-      ut.passes("test_empty passes");
+    ut.passes("test_empty passes");
   else
-      ut.failure("test_empty failed");
+    ut.failure("test_empty failed");
 
   return;
 }
@@ -186,9 +186,9 @@ int main(int argc, char *argv[]) {
     test_cyclic(ut);
     test_empty(ut);
     if (ut.numFails == 0)
-        ut.passes("All passed");
+      ut.passes("All passed");
     else
-        ut.failure("Failed");
+      ut.failure("Failed");
   }
   UT_EPILOG(ut);
 }
